Use new WinRT APIs for SpecialFolders
