+++ src/System.Runtime.Extensions/src/System.Runtime.Extensions.csproj
@@ -89,10 +89,21 @@
       <Link>Common\System\HResults.cs</Link>
     </Compile>
   </ItemGroup>
+  <ItemGroup Condition="'$(TargetsWindows)' == 'true' and ('$(TargetGroup)' == 'uapaot' or '$(TargetGroup)' == 'uap')">
+    <Compile Include="System\Environment.WinRT.cs" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetsWindows)' == 'true' and '$(TargetGroup)' != 'uapaot' and '$(TargetGroup)' != 'uap'">
+    <Compile Include="System\Environment.Win32.cs" />
+    <Compile Include="$(CommonPath)\Interop\Windows\secur32\Interop.GetUserNameExW.cs" >
+      <Link>Common\Interop\Windows\secur32\Interop.GetUserNameExW.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\Interop\Windows\shell32\Interop.SHGetKnownFolderPath.cs">
+      <Link>Common\Interop\Windows\shell32\Interop.SHGetKnownFolderPath.cs</Link>
+    </Compile>
+  </ItemGroup>
   <!-- WINDOWS: Shared CoreCLR and .NET Native -->
   <ItemGroup Condition="'$(TargetsWindows)' == 'true'">
     <Compile Include="System\Environment.Windows.cs" />
-    <Compile Include="System\Environment.Win32.cs"  Condition="'$(TargetGroup)' != 'uapaot' and '$(TargetGroup)' != 'uap'" />
     <Compile Include="System\Runtime\Versioning\VersioningHelper.Windows.cs" />
     <Compile Include="System\Diagnostics\Stopwatch.Windows.cs" />
     <Compile Include="$(CommonPath)\System\IO\Win32Marshal.cs">
@@ -160,12 +171,6 @@
     </Compile>
     <Compile Include="$(CommonPath)\Interop\Windows\kernel32\Interop.SYSTEM_INFO.cs">
       <Link>Common\Interop\Windows\kernel32\Interop.SYSTEM_INFO.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\Interop\Windows\secur32\Interop.GetUserNameExW.cs" Condition="'$(TargetGroup)' != 'uapaot' and '$(TargetGroup)' != 'uap'" >
-      <Link>Common\Interop\Windows\secur32\Interop.GetUserNameExW.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\Interop\Windows\shell32\Interop.SHGetKnownFolderPath.cs">
-      <Link>Common\Interop\Windows\shell32\Interop.SHGetKnownFolderPath.cs</Link>
     </Compile>
     <Compile Include="$(CommonPath)\System\IO\DriveInfoInternal.Win32.cs">
       <Link>Common\System\IO\DriveInfoInternal.Win32.cs</Link>
@@ -250,14 +255,24 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\System.Private.Uri\src\System.Private.Uri.csproj" />
-    <ProjectReference Include="..\..\System.Diagnostics.Debug\src\System.Diagnostics.Debug.csproj" />
-    <ProjectReference Include="..\..\System.Security.Principal\src\System.Security.Principal.csproj" />
-    <ProjectReference Include="..\..\System.Runtime\src\System.Runtime.csproj" />
+    <ProjectReference Include="..\..\System.Private.Uri\src\System.Private.Uri.csproj">
+        <AdditionalProperties>$(AdditionalProperties);TargetGroup=$(TargetGroup)</AdditionalProperties>
+    </ProjectReference>
+    <ProjectReference Include="..\..\System.Diagnostics.Debug\src\System.Diagnostics.Debug.csproj">
+        <AdditionalProperties>$(AdditionalProperties);TargetGroup=$(TargetGroup)</AdditionalProperties>
+    </ProjectReference>
+    <ProjectReference Include="..\..\System.Security.Principal\src\System.Security.Principal.csproj">
+        <AdditionalProperties>$(AdditionalProperties);TargetGroup=$(TargetGroup)</AdditionalProperties>
+    </ProjectReference>
+    <ProjectReference Include="..\..\System.Runtime\src\System.Runtime.csproj">
+        <AdditionalProperties>$(AdditionalProperties);TargetGroup=$(TargetGroup)</AdditionalProperties>
+    </ProjectReference>
   </ItemGroup>
-  <ItemGroup Condition="'$(TargetGroup)' == 'uapaot'">
+  <ItemGroup Condition="'$(TargetGroup)' == 'uapaot' or '$(TargetGroup)' == 'uap'">
     <Reference Include="mscorlib" />
     <Reference Include="Windows" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == 'uapaot'">
     <ReferenceFromRuntime Include="System.Private.Interop" />
   </ItemGroup>
   <ItemGroup>
